requirements.txt
Django==1.11
django-cors-headers==2.3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
MySQL-python==1.2.5
PyJWT==1.6.4
pytz==2018.5
